moved ItemsAccessor to other accessors
