--- src/main/java/teamport/aether/mixin/MinecraftMixin.java
@@ -1,18 +1,0 @@
-package teamport.aether.mixin;
-
-import net.minecraft.client.Minecraft;
-import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
-import teamport.aether.AetherClient;
-
-@Mixin(value = Minecraft.class, remap = false)
-
-public abstract class MinecraftMixin {
-    @Inject(method = "startGame", at = @At(value = "INVOKE", target = "Lnet/minecraft/core/achievement/stat/StatList;init()V"))
-    public void initStats(CallbackInfo ci) {
-        ItemsAccessor.invokeInitStats();
-        AetherClient.initAchievementsPage();
-    }
-}
